lab_2/backend/requirements.txt
# Web framework
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9

# Database
sqlalchemy>=2.0.30
alembic>=1.13.1
asyncpg>=0.29.0
pgvector>=0.3.0

# Config
pydantic-settings>=2.2.1

# NLP
spacy>=3.7.4

# Embeddings
sentence-transformers>=3.0.0

# File parsers
pypdf2>=3.0.1
python-docx>=1.1.2
striprtf>=0.0.26

# Utilities
aiofiles>=23.2.1
